finished

+++ includes/delete_img.inc.php
@@ -8,22 +8,27 @@ if(empty($_POST['delete']))
     exit();
 }
 else {
-    $image_Del = $_POST['delete'];
-    $fileName = "../images/" . $image_Del . "*";
-    $fileInfo = glob($fileName);
-    $fileExt = explode(".", $fileInfo[0]);
-    $fileActualExt = $fileExt[3];
+    $sqlCheck = "SELECT * FROM uploaded_img WHERE img_name='$_POST[delete]';";
+    $resultsCheck = $connexion->query($sqlCheck);
+    $row = $resultsCheck->fetch(PDO::FETCH_ASSOC);
+    if($row['user_id'] == $_SESSION['u_id']) {
+        $image_Del = $_POST['delete'];
+        $fileName = "../images/" . $image_Del . "*";
+        $fileInfo = glob($fileName);
+        $fileExt = explode(".", $fileInfo[0]);
+        $fileActualExt = $fileExt[3];
 
-    $fileActualName = $image_Del . "." . $fileActualExt;
-    $file = "../images/" . $image_Del . "." . $fileActualExt;
-    if (!unlink($file)) {
+        $fileActualName = $image_Del . "." . $fileActualExt;
+        $file = "../images/" . $image_Del . "." . $fileActualExt;
+        if (!unlink($file)) {
 
-        echo "File was not deleted";
-    } else {
-        $sql = "DELETE FROM uploaded_img WHERE img_name=?;";
-        $stmt = $connexion->prepare($sql);
-        $stmt->execute(array($fileActualName));
-        echo "File was deleted";
+            echo "File was not deleted";
+        } else {
+            $sql = "DELETE FROM uploaded_img WHERE img_name=?;";
+            $stmt = $connexion->prepare($sql);
+            $stmt->execute(array($fileActualName));
+            echo "File was deleted";
+        }
     }
     header("Location: ../img_galery.php");
 }
